refactor: split main page
