+++ src/routes/+page.svelte
@@ -1,341 +1,20 @@
 <script lang="ts">
-	import { onMount } from 'svelte';
-	import { Carousel } from 'flowbite-svelte';
-
-	let countdown = '';
-	const eventDate = new Date('2025-05-31T23:59:59');
-	// 地圖位置的 Google Maps 連結
-	const locations = {
-		activity:
-			'https://www.google.com/maps/embed?pb=!1m18!1m12!1m3!1d3615.774712013287!2d121.53454321193558!3d25.007770077740933!2m3!1f0!2f0!3f0!3m2!1i1024!2i768!4f13.1!3m3!1m2!1s0x3442aa1e27ab738b%3A0x2e5be38c37583475!2z5ZyL56uL6Ie654Gj5bir56-E5aSn5a24IOWFrOmkqOagoeWNgA!5e0!3m2!1szh-TW!2stw!4v1743730247104!5m2!1szh-TW!2stw',
-		dorm: 'https://www.google.com/maps/embed?pb=!1m18!1m12!1m3!1d3615.367900441113!2d121.52431157429362!3d25.021585877824418!2m3!1f0!2f0!3f0!3m2!1i1024!2i768!4f13.1!3m3!1m2!1s0x3442a9fd810cbc83%3A0x38bf4a0a52f72137!2z5om_5pSc6KGM5peFIC0g5Y-w5YyX5Y-w5aSn6aSo!5e0!3m2!1szh-TW!2stw!4v1743789653775!5m2!1szh-TW!2stw'
-	};
-
-	let currentMap = locations.activity; // 預設顯示活動地點
-
-	let stage; // 報名階段
-	const now = new Date();
-	if (now < new Date('2025-05-01T00:00:00')) {
-		stage = 0; // 早鳥報名
-	} else if (now < new Date('2025-05-20T00:00:00')) {
-		stage = 1; // 一階報名
-	} else if (now >= new Date('2025-05-27T00:00:00') && now < new Date('2025-06-02T00:00:00')) {
-		stage = 2; // 二階報名
-	} else if (now >= new Date('2025-06-02T00:00:00')) {
-		stage = 3; // 報名已截止
-	} else {
-		stage = 4; // 非報名時段
-	}
-
-	let stageText = ''; // 報名階段文字
-	let dateTexts = [
-		'2025/04/21 - 2025/04/30', // 早鳥報名截止日期
-		'2025/05/01 - 2025/05/19', // 一階報名截止日期
-		'2025/05/27 - 2025/06/01', // 二階報名截止日期
-		'', // 報名已截止
-		'(若有名額則 5/27 後開放二階報名)' // 非報名時段
-	];
-	if (stage === 0) {
-		stageText = '早鳥報名：' + dateTexts[stage];
-	} else if (stage === 1) {
-		stageText = '一階報名：' + dateTexts[stage];
-	} else if (stage === 2) {
-		stageText = '二階報名：' + dateTexts[stage];
-	} else if (stage === 3) {
-		stageText = '報名已截止' + dateTexts[stage];
-	} else if (stage === 4) {
-		stageText = '暫不受理報名 ' + dateTexts[stage];
-	}
-
-	let stageCostText = ''; // 報名費用文字
-	let originalCostText = ''; // 原始報名費用
-	if (stage === 0) {
-		originalCostText = '7300';
-		stageCostText = '6800 元'; // 早鳥報名費用
-	} else if (stage === 1) {
-		originalCostText = '';
-		stageCostText = '7300 元'; // 一階報名費用
-	} else if (stage === 2) {
-		originalCostText = '';
-		stageCostText = '7300 元'; // 二階報名費用
-	} else if (stage === 3) {
-		originalCostText = '';
-		stageCostText = '7300 元 (報名已截止)'; // 報名已截止
-	} else if (stage === 4) {
-		originalCostText = '';
-		stageCostText = '7300 元'; // 非報名時段
-	}
-
-	// 倒數計時邏輯
-	onMount(() => {
-		const interval = setInterval(() => {
-			const now = new Date();
-			const diff = eventDate.getTime() - now.getTime();
-
-			if (diff <= 0) {
-				clearInterval(interval);
-				countdown = '報名已截止';
-			} else {
-				const days = Math.floor(diff / (1000 * 60 * 60 * 24));
-				const hours = Math.floor((diff / (1000 * 60 * 60)) % 24);
-				const minutes = Math.floor((diff / (1000 * 60)) % 60);
-				const seconds = Math.floor((diff / 1000) % 60);
-				countdown = `${days} 天 ${hours} 小時 ${minutes} 分鐘 ${seconds} 秒`;
-			}
-		}, 1000);
-	});
-
-	// 切換地圖位置
-	function switchLocation(location: 'activity' | 'dorm') {
-		currentMap = locations[location];
-	}
-
-	const imageFiles = import.meta.glob('/src/images/*.{jpg,jpeg,png,gif}', { eager: true });
-	let images = Object.entries(imageFiles).map(([path, module], index) => ({
-		src: path,
-		alt: `Image ${index + 1}`
-	}));
-
-	// 計算navbar高度
-	let navbarHeight = 0;
-	onMount(() => {
-		const navbar = document.querySelector('nav');
-		if (navbar) {
-			navbarHeight = navbar.offsetHeight;
-		}
-	});
-
-	// 交通方式
-	let activeTab = 'Bus';
+	import HeroPage from '../sections/HeroPage.svelte';
+	import Introduction from '../sections/Introduction.svelte';
+	import CampInfo from '../sections/CampInfo.svelte';
+	import Transportation from '../sections/Transportation.svelte';
+	import Hackathon from '../sections/Hackathon.svelte';
+	import CampHistory from '../sections/CampHistory.svelte';
 </script>
 
-<section class="flex items-center px-8" style="height: calc(100vh - {navbarHeight}px);">
-	<!-- 左側圖片 -->
-	<div class="flex flex-1 items-center justify-center">
-		<img src="cube.png" alt="資工營主視覺" class="h-auto w-full" />
-	</div>
+<HeroPage/>
 
-	<!-- 右側文字區塊 -->
-	<div class="flex flex-1 flex-col items-start justify-center space-y-4 pl-6">
-		<h1 class="text-3xl">2025 臺師大資工營</h1>
-		<h2 class="text-6xl/20">資遊你和我的師界</h2>
-		<p class="text-3xl">
-			報名期間
-			<span class="text-6xl">04/21 - 05/31</span>
-		</p>
-		<a
-			href="https://forms.gle/tZzXBLQuiR8QSpdKA"
-			class="mt-4 rounded-none bg-yellow-300 px-10 py-3 text-3xl font-bold text-black transition hover:bg-yellow-500"
-		>
-			立即報名
-		</a>
-	</div>
-</section>
+<Introduction/>
 
-<section class="mx-18 mb-20 min-w-[720px] text-pretty">
-	<h1 class="mb-10 text-center text-6xl">
-		今年夏天<br />（史詩級的音效）
-	</h1>
+<CampInfo/>
 
-	<div class="text-center text-2xl/12">
-		「各位公測玩家們準備好了嗎？」<br />
-		「Yooooooooooooooo！」<br />
-		由《你在”code“什麼》團隊攜手無名小卒打造。 <br />
-		耗資我也不知道。（滑倒） <br />
-		「打敗那隻 Boss！！！」（登登） <br />
-		《資遊你和我的師界》！（哐當）<br />
-		「所以說，Boss 人咧？」<br />
-		「欸？！！！」<br />
-		「讓我們一起進入遊戲的世界吧！」<br />
-		今年暑假，我們不見不散！（收尾的音效）<br /><br />
-		伊莎貝拉：「非哥，你在”code“什～麼？」<br />
-		非哥：「四月二十一，早鳥報名開始！」<br /><br />
-	</div>
-</section>
+<Transportation/>
 
-<section class="mx-18 mb-20 min-w-[720px] text-pretty">
-	<div class="my-3">
-		<span class="text-4xl">營隊資訊</span>
-	</div>
-	<div class="border-3 border-white bg-black/40">
-		<p class="border-b-3 border-white p-7 text-center text-3xl">
-			距離報名截止剩餘：{countdown}
-		</p>
-		<div class="m-7 grid w-[95%] grid-cols-[15%_85%] items-center justify-center gap-x-6 gap-y-6">
-			<div class="border-r-3 py-1 text-center text-3xl">報名時間</div>
-			<div class="justify-self-start text-2xl">
-				{stageText}
-			</div>
+<Hackathon/>
 
-			<div class="border-r-3 py-1 text-center text-3xl">報名費用</div>
-			<div class="justify-self-start text-2xl">
-				<span class="text-xl text-red-500 line-through">{originalCostText}</span>
-				{stageCostText}
-			</div>
-
-			<div class="border-r-3 py-1 text-center text-3xl">活動時間</div>
-			<div class="justify-self-start text-2xl">7/1 (二) - 7/4 (五)</div>
-
-			<div class="border-r-3 py-1 text-center text-3xl">詳細資訊</div>
-			<div class="justify-self-start text-2xl">
-				<a
-					href="https://docs.google.com/document/d/1yZE3Mumvc52DEaeNavlcUavxazWvIfe8/edit?rtpof=true&sd=true"
-					class="hover:text-yellow-300">點此查看 活動簡章</a
-				>
-			</div>
-
-			<button
-				type="button"
-				class="border-3 border-white py-3 text-center text-3xl hover:cursor-pointer hover:bg-yellow-300 hover:text-black"
-				on:click={() => switchLocation('activity')}
-				on:keydown={(e) => e.key === 'Enter' && switchLocation('activity')}
-			>
-				活動地點
-			</button>
-			<span class="justify-self-start text-2xl"> 國立台灣師範大學公館校區 </span>
-
-			<button
-				type="button"
-				class="border-3 border-white py-3 text-center text-3xl hover:cursor-pointer hover:bg-yellow-300 hover:text-black"
-				on:click={() => switchLocation('dorm')}
-				on:keydown={(e) => e.key === 'Enter' && switchLocation('dorm')}
-			>
-				住宿地點
-			</button>
-			<span class="justify-self-start text-2xl">
-				承攜行旅-台北台大館<br />（台北市中正區羅斯福路三段 98 號）
-			</span>
-		</div>
-
-		<div class="mx-auto my-5 aspect-[752/345] w-[90%] overflow-hidden rounded-lg">
-			<iframe
-				title="Location Map"
-				src={currentMap}
-				class="h-full w-full border-0"
-				allowfullscreen
-				loading="lazy"
-			>
-			</iframe>
-		</div>
-	</div>
-</section>
-
-<section class="mx-18 mb-20 min-w-[720px] text-pretty">
-	<div class="my-3">
-		<span class="text-4xl text-white">交通方式</span>
-	</div>
-	<div class="border-3 border-white bg-black/40">
-		<div class="border-b-3 border-white text-2xl/12 flex">
-            <button 
-                class="flex-1 hover:cursor-pointer
-                {activeTab === 'Bus' ? 'bg-yellow-300 text-black' : ''}" 
-                on:click={() => {activeTab = 'Bus'}}
-            >
-            搭公車
-            </button>
-            <div class="w-[3px] bg-white"></div>
-			<button 
-                class="flex-1 hover:cursor-pointer
-                {activeTab === 'MRT' ? 'bg-yellow-300 text-black' : ''}" 
-                on:click={() => {activeTab = 'MRT'}}
-            >
-            搭捷運
-            </button>
-		</div>
-		<div class="text-2xl p-5 w-full text-center">
-            {#if activeTab == 'Bus'}
-            <p>搭乘</p>
-            <p class="my-3 text-xl"> 
-                0 南、109、236 區、251、251 區、252、<br>
-                253、278、278 區、280、280 直、284、<br>
-                505、52、530、606、643、644、<br>
-                648、660、668、棕 11、藍 28、<br>
-                羅斯福路幹線、松江新生幹線、復興幹線、基隆路幹線、景美-榮總（快）</p>
-            <p>至「師大分部」站</p>
-            <p class="text-5xl my-5">↓</p>
-            <p>向西步行約 2 分鐘至國立台灣師範大學公館校區門口</p>
-
-            {:else}
-            <p>搭乘松山新店線（綠線）至「公館」站 1 號出口</p>
-            <p class="text-5xl my-5">↓</p>
-            <p>向東南步行約 12 分鐘至國立台灣師範大學公館校區門口</p>
-
-            {/if}
-		</div>
-	</div>
-</section>
-
-<section class="mx-18 mb-20 min-w-[720px] text-pretty">
-	<div class="my-3">
-		<span class="text-4xl">黑客松</span>
-	</div>
-	<div class="border-3 border-white bg-black/40 p-5 text-center text-2xl/12">
-		<p>
-			在上了這麼多 Discord Bot 的課，你是不是也開始想問：<br />
-			「學這些到底可以幹嘛啊？」<br />
-			在營隊的最後，我們要來點刺激的：<br />
-			💥 每組都要寫出一隻超酷、超實用、超創意的 Discord Bot！<br />
-			不但有機會拿獎品，還有神秘的好吃點心獎勵，<br />
-			獻給那些熬夜寫 code、debug 到天亮（希望不會啦 😅）的你！<br />
-			等等？你問主題是什麼？<br />
-			嘿嘿～我們的主題超級多元，<br />
-			有的畫畫、有的給資源、有的像小遊戲一樣超上癮～<br />
-			📌 畫畫機器人？<br />
-			📌 養成遊戲機器人？<br />
-			📌 學習資源機器人？<br />
-			什麼都有，只有你想不到，沒有我們不敢讓你挑戰的！<br />
-			想知道我們黑客松主題到底是什們呢？<br />
-			那就來參加《2025 師大資工營》來尋找答案吧！<br />
-			用一場最狂的 Discord Bot 黑客松，<br />
-			為你的高中生活畫下一個印象深刻的驚嘆號吧！🔥<br />
-		</p>
-	</div>
-</section>
-
-<section class="mx-18 mb-20 min-w-[720px] text-pretty">
-	<div class="my-3">
-		<span class="text-4xl">歷屆回顧</span>
-	</div>
-	<div class="border-3 border-white bg-black/40 text-center">
-		<div class="border-b-3 border-white text-2xl/15">2024 | 你們在code什麼</div>
-		<div class="group">
-			<Carousel
-				{images}
-				duration={5000}
-				let:Controls
-				let:Indicators
-				class="aspect-[352/181] rounded-none"
-				style="height:100%;"
-			>
-				<Indicators />
-				<Controls let:ControlButton let:changeSlide>
-					<ControlButton
-						name="Previous"
-						forward={false}
-						on:click={() => changeSlide(false)}
-						class=" z-0 translate-x-5 opacity-0 transition-opacity duration-500 group-hover:opacity-100"
-					>
-						<div
-							class="flex h-12 w-12 items-center justify-center rounded-full bg-black/40 transition-colors hover:cursor-pointer hover:bg-black/80"
-						>
-							<img src="/icons/preview.png" alt="preview" class="h-3/5 w-3/5 object-contain" />
-						</div>
-					</ControlButton>
-
-					<ControlButton
-						name="Next"
-						forward={true}
-						on:click={() => changeSlide(true)}
-						class="z-0 -translate-x-5 opacity-50 transition-opacity duration-500 group-hover:opacity-100"
-					>
-						<div
-							class="flex h-12 w-12 items-center justify-center rounded-full bg-black/40 transition-colors hover:cursor-pointer hover:bg-black/80"
-						>
-							<img src="/icons/next.png" alt="preview" class="h-3/5 w-3/5 object-contain" />
-						</div>
-					</ControlButton>
-				</Controls>
-			</Carousel>
-		</div>
-	</div>
-</section>
+<CampHistory/>
