+++ src/sections/Hackathon.svelte
@@ -1,0 +1,26 @@
+<section class="mx-18 mb-20 min-w-[720px] text-pretty">
+	<div class="my-3">
+		<span class="text-4xl">黑客松</span>
+	</div>
+	<div class="border-3 border-white bg-black/40 p-5 text-center text-2xl/12">
+		<p>
+			在上了這麼多 Discord Bot 的課，你是不是也開始想問：<br />
+			「學這些到底可以幹嘛啊？」<br />
+			在營隊的最後，我們要來點刺激的：<br />
+			💥 每組都要寫出一隻超酷、超實用、超創意的 Discord Bot！<br />
+			不但有機會拿獎品，還有神秘的好吃點心獎勵，<br />
+			獻給那些熬夜寫 code、debug 到天亮（希望不會啦 😅）的你！<br />
+			等等？你問主題是什麼？<br />
+			嘿嘿～我們的主題超級多元，<br />
+			有的畫畫、有的給資源、有的像小遊戲一樣超上癮～<br />
+			📌 畫畫機器人？<br />
+			📌 養成遊戲機器人？<br />
+			📌 學習資源機器人？<br />
+			什麼都有，只有你想不到，沒有我們不敢讓你挑戰的！<br />
+			想知道我們黑客松主題到底是什們呢？<br />
+			那就來參加《2025 師大資工營》來尋找答案吧！<br />
+			用一場最狂的 Discord Bot 黑客松，<br />
+			為你的高中生活畫下一個印象深刻的驚嘆號吧！🔥<br />
+		</p>
+	</div>
+</section>
